feat: Create default resource support
In order to avoid loading the GLSL shaders synchronously, the files
must be loaded before the SimpleShader object. This should take place
in the enging initialization. Also, the default resource support can
create a sharable instance of the SimpleShader. This resource can support
all loading and sharing of future engine resources.